Set of functions in example RcppEigen package

These four functions are created when RcppEigen.package.skeleton() is invoked to create a skeleton packages.

rcppeigen_hello_world()
rcppeigen_outerproduct(x)
rcppeigen_innerproduct(x)
rcppeigen_bothproducts(x)

Arguments

x
a numeric vector

Value

rcppeigen_hello_world() does not return a value, but displays a message to the console.

rcppeigen_outerproduct() returns a numeric matrix computed as the outer (vector) product of x.

rcppeigen_innerproduct() returns a double computer as the inner (vector) product of x.

rcppeigen_bothproducts() returns a list with both the outer and inner products.

Details

These are example functions which should be largely self-explanatory. Their main benefit is to demonstrate how to write a function using the Eigen C++ classes, and to have to such a function accessible from R.

References

See the documentation for Eigen, and RcppEigen, for more details.

Examples

  x <- sqrt(1:4)
  rcppeigen_innerproduct(x)

#> Error in rcppeigen_innerproduct(x): could not find function "rcppeigen_innerproduct"

  rcppeigen_outerproduct(x)

#> Error in rcppeigen_outerproduct(x): could not find function "rcppeigen_outerproduct"